Makefile: support static-only builds
One of the problems with udev is that systemd cannot be built
statically[1].  With libudev-zero, on the other hand, no code changes
are required to be able to do a static-only build, only Makefile
changes.

With this change, it's possible to "make install-static" to do a
static-only build of libudev-zero.

[1]: https://github.com/systemd/systemd/pull/20621#issuecomment-912014839